Update tests in test_schema.py
