+++ tests/schemas/transform_model.yaml
@@ -1,0 +1,18 @@
+$schema: http://stsci.edu/schemas/yaml-schema/draft-01
+id: http://example.com/schemas/transform_model
+title: Model that includes a transform.
+
+allOf:
+  - $ref: core_metadata
+  - type: object
+    properties:
+      meta:
+        type: object
+        properties:
+          origin:
+            type: string
+          telescope:
+            type: string
+      transform:
+        $ref: http://stsci.edu/schemas/asdf/transform/transform-1.2.0
+...
